delete extra files and stuff
