--- src/combine_players.json
@@ -1,18 +1,0 @@
-{
-	"Darrell": [
-		"https://www.pgstats.com/melee/player/Darrell?id=S211440",
-		"https://www.pgstats.com/melee/player/Darrell?id=S4651"
-	],
-	"Rainy": [
-		"https://www.pgstats.com/melee/player/rainy?id=S25939",
-		"https://www.pgstats.com/melee/player/rainy?id=S3275225"
-	],
-	"SpacePigeon": [
-		"https://www.pgstats.com/melee/player/SpacePigeon?id=S24024",
-		"https://www.pgstats.com/melee/player/SpacePigeon?id=S2621428"
-	],
-	"Nickemwit": [
-		"https://www.pgstats.com/melee/player/Nickemwit?id=S3355",
-		"https://www.pgstats.com/melee/player/Nickemwit?id=S3251331"
-	]
-}
