--- src/good-players.csv
@@ -1,39 +1,0 @@
-Player,pgstats
-Kevbot,https://www.pgstats.com/melee/player/Kevbot?id=S12293
-Iron,https://www.pgstats.com/melee/player/Iron?id=S897692
-Aerius,https://www.pgstats.com/melee/player/Aerius?id=S155114
-Arcadia,https://www.pgstats.com/melee/player/Arcadia?id=S3772
-Azel,https://www.pgstats.com/melee/player/Azel?id=S68261
-BigDK,https://www.pgstats.com/melee/player/BigDK?id=S48685
-Biscuito,https://www.pgstats.com/melee/player/biscuito?id=S994409
-blargh257,https://www.pgstats.com/melee/player/blargh257?id=S11126
-Darkatma,https://www.pgstats.com/melee/player/Darkatma?id=S7162
-darkwizard123,https://www.pgstats.com/melee/player/darkwizard123?id=S22848
-Depp,https://www.pgstats.com/melee/player/Depp?id=S27284
-FreeHops,https://www.pgstats.com/melee/player/FreeHops?id=S1077796
-MOJOE,https://www.pgstats.com/melee/player/MOJOE?id=S14354
-Pancakes,https://www.pgstats.com/melee/player/Pancakes?id=S47324
-PRZ,https://www.pgstats.com/melee/player/PRZ?id=S16361
-rainy1,https://www.pgstats.com/melee/player/rainy?id=S25939
-rainy2,https://www.pgstats.com/melee/player/rainy?id=S3275225
-Ralph,https://www.pgstats.com/melee/player/Ralph?id=S4505
-redpandamaster,https://www.pgstats.com/melee/player/redpandamaster?id=S25916
-rom,https://www.pgstats.com/melee/player/rom?id=S3899
-S2J,https://www.pgstats.com/melee/player/S2J?id=S1017
-SFAT,https://www.pgstats.com/melee/player/SFAT?id=S1019
-Shroomed,https://www.pgstats.com/melee/player/Shroomed?id=S1013
-Snap,https://www.pgstats.com/melee/player/Snap?id=S3377
-SpacePigeon1,https://www.pgstats.com/melee/player/SpacePigeon?id=S24024
-SpacePigeon2,https://www.pgstats.com/melee/player/SpacePigeon?id=S2621428
-Spark,https://www.pgstats.com/melee/player/Spark?id=S3357
-Typhoon,https://www.pgstats.com/melee/player/Typhoon?id=S221423
-Umarth,https://www.pgstats.com/melee/player/Umarth?id=S15504
-vesper,https://www.pgstats.com/melee/player/vesper?id=S1263697
-NathanSandwich,https://www.pgstats.com/melee/player/NathanSandwich?id=S4176
-Tucan,https://www.pgstats.com/melee/player/DiplomaticTucan?id=S13005
-DR Grin,https://www.pgstats.com/melee/player/Dr.%20Grin?id=S4259
-BQ,https://www.pgstats.com/melee/player/BQ?id=S319042
-Darrell1,https://www.pgstats.com/melee/player/Darrell?id=S211440
-Darrell2,https://www.pgstats.com/melee/player/Darrell?id=S4651
-Rocky,https://www.pgstats.com/melee/player/Rocky?id=S4215
-20k lions(Tucan),https://www.pgstats.com/melee/player/20%2C000%20Lions?id=S4317
